# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/patrocinios": {
            "get": {
                "summary": "Get list of patrocinios",
                "operationId": "210c9aab30ea4c0d455d2a15fe4d3eb7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new patrocinio",
                "operationId": "49644b7f00148d9a5bcfd6c3cbcd2cdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "EquipoId",
                                    "PatrocinadorId",
                                    "FechaInicio",
                                    "FechaFin"
                                ],
                                "properties": {
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "PatrocinadorId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "FechaInicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "FechaFin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/patrocinios/{id}": {
            "get": {
                "summary": "Get patrocinio by ID",
                "operationId": "1b61622ccc44908090b2d5aa324b9eb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update patrocinio by ID",
                "operationId": "bba302101499d7e4f29a40d90d9b94f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "EquipoId",
                                    "PatrocinadorId",
                                    "FechaInicio",
                                    "FechaFin"
                                ],
                                "properties": {
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "PatrocinadorId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "FechaInicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "FechaFin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete patrocinio by ID",
                "operationId": "b1d6085d7bfe36772297bc0cf07af029",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/personal": {
            "get": {
                "summary": "Get list of personal",
                "operationId": "b06eb1dcbe33c370fd8757aa138dbec7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new personal",
                "operationId": "af3f4872a7d605eba5bb92df50691518",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "RolId",
                                    "EquipoId"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "RolId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/personal/{id}": {
            "get": {
                "summary": "Get personal by ID",
                "operationId": "88bd006eb75fd8592457a34b363e0c77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update personal by ID",
                "operationId": "c1591b29f8695dd29d077f815e3304eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "RolId",
                                    "EquipoId"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "RolId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete personal by ID",
                "operationId": "7a6119e00ca786bb540f3c9c6787026b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}